moving default methods to abstract class
